+++ Changelog.md
@@ -3,7 +3,10 @@
 ## Unreleased
 
 * motoko (`moc`)
-  * Experimental support for Mixins (#5459)
+
+  * Experimental support for `implicit` argument declarations (#5517).
+
+  * Experimental support for Mixins (#5459).
 
 ## 0.16.3 (2025-09-29)